requirements.txt
altgraph==0.17.4
auto-py-to-exe==2.48.1
beautifulsoup4==4.14.2
bottle==0.13.4
bottle-websocket==0.2.9
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
Eel==0.18.2
future==1.0.0
gevent==25.9.1
gevent-websocket==0.10.1
greenlet==3.2.4
idna==3.11
importlib_resources==6.5.2
ordered-set==4.1.0
packaging==25.0
pefile==2023.2.7
pillow==12.0.0
playwright==1.56.0
pycparser==2.23
pyee==13.0.0
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pyparsing==3.2.5
pywin32-ctypes==0.2.3
reportlab==4.4.4
requests==2.32.5
setuptools==80.9.0
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.5.0
zope.event==6.0
zope.interface==8.0.1
zstandard==0.25.0
